+++ src/rtp/rtp_codec.h
@@ -21,27 +21,28 @@ class RtpCodec {
   ~RtpCodec();
 
  public:
-  void Encode(uint8_t* buffer, size_t size, std::vector<RtpPacket>& packets);
-  void Encode(VideoFrameType frame_type, uint8_t* buffer, size_t size,
+  void Encode(uint8_t* buffer, uint32_t size, std::vector<RtpPacket>& packets);
+  void Encode(VideoFrameType frame_type, uint8_t* buffer, uint32_t size,
               std::vector<RtpPacket>& packets);
   size_t Decode(RtpPacket& packet, uint8_t* payload);
 
   //  protected:
   //   void OnReceiveFrame(uint8_t* payload) = 0;
  private:
-  bool IsKeyFrame(const uint8_t* buffer, size_t size);
+  bool IsKeyFrame(const uint8_t* buffer, uint32_t size);
 
-  void EncodeAv1(uint8_t* buffer, size_t size, std::vector<RtpPacket>& packets);
+  void EncodeAv1(uint8_t* buffer, uint32_t size,
+                 std::vector<RtpPacket>& packets);
 
  private:
-  uint32_t version_ = 0;
+  uint8_t version_ = 0;
   bool has_padding_ = false;
   bool has_extension_ = false;
   uint32_t total_csrc_number_ = 0;
   bool marker_ = false;
   uint32_t payload_type_ = 0;
   uint16_t sequence_number_ = 0;
-  uint32_t timestamp_ = 0;
+  uint64_t timestamp_ = 0;
   uint32_t ssrc_ = 0;
   std::vector<uint32_t> csrcs_;
   uint16_t profile_ = 0;
